PA-209 rename address_book_model.py
